+++ BeatTogether.DedicatedServer.Interface/BeatTogether.DedicatedServer.Interface.csproj
@@ -7,7 +7,7 @@
     <Authors>pythonology,cubic</Authors>
     <Company>pythonology,cubic</Company>
     <RepositoryUrl>https://github.com/pythonology/BeatTogether.DedicatedServer</RepositoryUrl>
-    <Version>1.3.4</Version>
+    <Version>1.3.6</Version>
     <Nullable>enable</Nullable>
   </PropertyGroup>
