--- BeatTogether.DedicatedServer.Interface/DedicatedServerServiceContract.cs
@@ -1,15 +1,0 @@
-﻿using Autobus;
-using BeatTogether.DedicatedServer.Interface.Events;
-
-namespace BeatTogether.DedicatedServer.Interface
-{
-    public class DedicatedServerServiceContract : BaseServiceContract
-    {
-        public override void Build(IServiceContractBuilder builder) =>
-            builder
-                .UseName("DedicatedServer")
-                .AddInterface<IMatchmakingService>()
-                //.AddEvent<MatchmakingServerStartedEvent>()
-                .AddEvent<MatchmakingServerStoppedEvent>();
-    }
-}
