+++ BeatTogether.DedicatedServer.Node/BeatTogether.DedicatedServer.Node.csproj
@@ -13,7 +13,7 @@
 
   <ItemGroup>
 	<PackageReference Include="BeatTogether.Extensions.Autobus.RabbitMQ" Version="1.0.2" />
-	<PackageReference Include="BeatTogether.MasterServer.Interface" Version="1.1.8" />	
+	<PackageReference Include="BeatTogether.MasterServer.Interface" Version="1.1.11" />	
   </ItemGroup>
 	
   <ItemGroup>
